+++ SummaryTask4/src/ua/nure/miroshnichenko/summarytask4/service/BanningServiceImpl.java
@@ -23,7 +23,7 @@ class BanningServiceImpl implements BanningService {
 		try {
 			DAO<User> userDao = factoryDAO.getUserDAO();
 			User user2 = userDao.find(id);
-			user2.setBanned((byte) 1);
+			user2.setBanned(1);
 			boolean result = userDao.update(user2);
 
 			return result;
